requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.18
Django==1.10
django-allauth==0.31.0
djangorestframework==3.9.1
kombu==3.0.37
LEPL==5.1.3
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
redis==2.10.3
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
